requirements.txt
langchain>=0.3.20,<1.0.0
langchain-core>=0.3.41,<1.0.0
langsmith>=0.1.125,<0.4



# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai



# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn
sentencepiece

fastapi
uvicorn 
transformers
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz